Allow r/configure to build cpp libs if they're found at ../cpp

+++ r/tools/linuxlibs.R
@@ -27,12 +27,21 @@ if (!file.exists(paste0(dst_dir, "/include/arrow/api.h"))) {
       cat(sprintf("*** %s does not exist; build will fail\n", localfile))
     }
     file.copy(localfile, "lib.zip")
+  } else if (file.exists("../cpp/src/arrow/api.h")) {
+    # We're in a git checkout of arrow. Build the C++ library
+    cat("*** Building C++ libraries\n")
+    # We'll need to compile R bindings with these libs, so delete any .o files
+    system("rm src/*.o")
+    system(sprintf("SOURCE_DIR=../cpp BUILD_DIR=libarrow/build DEST_DIR=%s inst/build_arrow_static.sh", dst_dir))
   } else {
     # Download static arrow
     # TODO: this should vary by distro/version
     # try(download.file("https://dl.bintray.com/nealrichardson/pyarrow-dev/libarrow.zip", "lib.zip", quiet = TRUE), silent = TRUE)
   }
-  dir.create(dst_dir, showWarnings = FALSE, recursive = TRUE)
-  unzip("lib.zip", exdir = dst_dir)
-  unlink("lib.zip")
+  # Finish up for the cases where we got a zip file of the libs
+  if (file.exists("lib.zip")) {
+    dir.create(dst_dir, showWarnings = FALSE, recursive = TRUE)
+    unzip("lib.zip", exdir = dst_dir)
+    unlink("lib.zip")
+  }
 }
